@@ -12,7 +12,7 @@ start = "src.main:start"
 test = "tests.conftest:run"
 
 [tool.poetry.dependencies]
-python = "3.9.6"
+python = "3.12.2"
 fastapi = "0.110.0"
 uvicorn = {extras = ["standard"], version = "^0.29.0"}
 sqlalchemy = "2.0.29"
@@ -31,6 +31,7 @@ python-multipart = "0.0.9"
 redis = "5.0.3"
 black = "24.3.0"
 pydantic-settings = "^2.2.1"
+pre-commit = "^3.7.0"
 
 [tool.poetry.group.test.dependencies]
 pytest = "8.1.1"
