@@ -1,5 +1,3 @@
-from typing import Union
-
 from sqlalchemy import select
 
 from src.database import async_session
@@ -10,7 +8,7 @@ from src.users.utils import Hasher
 
 class AuthService:
     @classmethod
-    async def get_user(cls, username: str, password: str) -> Union[SUser, None]:
+    async def get_user(cls, username: str, password: str) -> SUser | None:
         async with async_session() as session:
             async with session.begin():
                 query = select(UserOrm).where(UserOrm.username == username)
